Ensure streaming works in both prod and dev
Strict mode was making some weird things happen by making the chat stream run twice. The `responses` value seems to be double updating, once for each time `append` is called. I thought the `stop()` call in the teardown would prevent the first run from doing anything but apparently not. Also the `stop()` in the useEffect used to process the response content was making things work in strict mode, but not in production. I'm not great with react and I'm in use effect hell but this seems to fix things without outright turning off strict mode